requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.3.0
autocommand==2.2.2
av==14.0.1
babel==2.16.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
black==24.10.0
bleach==6.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cheroot==10.0.1
CherryPy==18.10.0
chumpy==0.70
click==8.1.8
cloudpathlib==0.20.0
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
Cython==3.0.11
cython_bbox==0.1.5
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.19.3
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.17.0
fonttools==4.55.4
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.25.2
hydra-colorlog==1.2.0
hydra-core==1.3.0
hydra_zen==0.13.0
idna==3.10
imageio==2.34.1
importlib_metadata==8.6.1
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lapx==0.5.11.post1
lazy_loader==0.4
lightning==2.3.0
lightning-utilities==0.11.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.11.0.86
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
portalocker==3.1.1
portend==3.2.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytorch-lightning==2.5.0.post0
pytorch3d @ https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py310_cu121_pyt230/pytorch3d-0.7.6-cp310-cp310-linux_x86_64.whl#sha256=e2845e8317768ed590b5999457879397e4876330b79bf01c2966c9e9dc116b87
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rtree==1.4.0
safetensors==0.5.2
scikit-image==0.24.0
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.20.0
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
smplx==0.1.28
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tempora==5.8.0
tensorboardX==2.6.2.2
termcolor==2.5.0
terminado==0.18.1
thinc==8.2.5
tifffile==2025.1.10
timm==0.9.12
tinycss2==1.4.0
tokenizers==0.13.3
tomli==2.2.1
torch==2.3.0+cu121
torchmetrics==1.6.1
torchvision==0.18.0+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.31.0
transforms3d==0.4.2
trimesh==4.6.0
triton==2.3.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.2.42
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.3.0
wandb==0.19.4
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wis3d==1.0.1
wrapt==1.17.2
yacs==0.1.8
yarl==1.18.3
zc.lockfile==3.0.post1
zipp==3.21.0
